requirements.txt
annotated-types==0.7.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
dateparser==1.2.1
Deprecated==1.2.18
freezegun==1.5.1
idna==3.10
isodate==0.7.2
keboola.component==1.6.10
keboola.http-client==1.0.1
keboola.utils==1.1.0
lxml==5.3.1
mock==5.1.0
platformdirs==4.3.6
pydantic==2.10.6
pydantic_core==2.27.2
pygelf==0.4.2
python-dateutil==2.9.0.post0
pytz==2025.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
six==1.17.0
typing_extensions==4.12.2
tzlocal==5.3
urllib3==2.3.0
wrapt==1.17.2
zeep==4.3.1
httpx==0.28.1
pytest==8.3.5